agent_results/end2end_gemini/iclr2025_question/claude_code/requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.31.0
datasets>=2.13.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
tensorboard>=2.13.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
sklearn>=0.0.post7
scikit-learn>=1.3.0
wikipedia-api>=0.6.0
pytrec-eval>=0.5.0
langchain>=0.0.267
chroma>=0.2.0

# For API access to closed-source LLMs
openai>=1.1.0
anthropic>=0.5.0

# For experiment tracking
wandb>=0.15.0
